+++ nilearn/decoding/decoder.py
@@ -298,16 +298,18 @@ class _BaseDecoder(LinearRegression, CacheMixin):
     %(low_pass)s
     %(high_pass)s
     %(t_r)s
-    mask_strategy: {'background' or 'epi'}, optional. Default: 'background'
-        The strategy used to compute the mask: use 'background' if your
-        images present a clear homogeneous background, and 'epi' if they
-        are raw EPI images. Depending on this value, the mask will be
-        computed from masking.compute_background_mask or
-        masking.compute_epi_mask.
+    %(mask_strategy)s
 
         .. note::
             This parameter will be ignored if a mask image is provided.
 
+        .. note::
+            Depending on this value, the mask will be computed from
+            :func:`nilearn.masking.compute_background_mask`,
+            :func:`nilearn.masking.compute_epi_mask`, or
+            :func:`nilearn.masking.compute_brain_mask`.
+
+        Default is 'background'.
     %(memory)s
     %(memory_level)s
     %(n_jobs)s
@@ -837,16 +839,18 @@ class Decoder(_BaseDecoder):
     %(low_pass)s
     %(high_pass)s
     %(t_r)s
-    mask_strategy: {'background' or 'epi'}, optional. Default: 'background'
-        The strategy used to compute the mask: use 'background' if your
-        images present a clear homogeneous background, and 'epi' if they
-        are raw EPI images. Depending on this value, the mask will be
-        computed from masking.compute_background_mask or
-        masking.compute_epi_mask.
+    %(mask_strategy)s
 
         .. note::
             This parameter will be ignored if a mask image is provided.
 
+        .. note::
+            Depending on this value, the mask will be computed from
+            :func:`nilearn.masking.compute_background_mask`,
+            :func:`nilearn.masking.compute_epi_mask`, or
+            :func:`nilearn.masking.compute_brain_mask`.
+
+        Default is 'background'.
     %(memory)s
     %(memory_level)s
     %(n_jobs)s
@@ -943,16 +947,18 @@ class DecoderRegressor(_BaseDecoder):
     %(low_pass)s
     %(high_pass)s
     %(t_r)s
-    mask_strategy: {'background' or 'epi'}, optional. Default: 'background'
-        The strategy used to compute the mask: use 'background' if your
-        images present a clear homogeneous background, and 'epi' if they
-        are raw EPI images. Depending on this value, the mask will be
-        computed from masking.compute_background_mask or
-        masking.compute_epi_mask.
+    %(mask_strategy)s
 
         .. note::
             This parameter will be ignored if a mask image is provided.
 
+        .. note::
+            Depending on this value, the mask will be computed from
+            :func:`nilearn.masking.compute_background_mask`,
+            :func:`nilearn.masking.compute_epi_mask`, or
+            :func:`nilearn.masking.compute_brain_mask`.
+
+        Default is 'background'.
     %(memory)s
     %(memory_level)s
     %(n_jobs)s
@@ -1056,16 +1062,18 @@ class FREMRegressor(_BaseDecoder):
     %(low_pass)s
     %(high_pass)s
     %(t_r)s
-    mask_strategy : {'background' or 'epi'}, optional. Default: 'background'
-        The strategy used to compute the mask: use 'background' if your
-        images present a clear homogeneous background, and 'epi' if they
-        are raw EPI images. Depending on this value, the mask will be
-        computed from masking.compute_background_mask or
-        masking.compute_epi_mask.
+    %(mask_strategy)s
 
         .. note::
             This parameter will be ignored if a mask image is provided.
 
+        .. note::
+            Depending on this value, the mask will be computed from
+            :func:`nilearn.masking.compute_background_mask`,
+            :func:`nilearn.masking.compute_epi_mask`, or
+            :func:`nilearn.masking.compute_brain_mask`.
+
+        Default is 'background'.
     %(memory)s
     %(memory_level)s
     %(n_jobs)s
@@ -1177,16 +1185,18 @@ class FREMClassifier(_BaseDecoder):
     %(low_pass)s
     %(high_pass)s
     %(t_r)s
-    mask_strategy : {'background' or 'epi'}, optional. Default: 'background'
-        The strategy used to compute the mask: use 'background' if your
-        images present a clear homogeneous background, and 'epi' if they
-        are raw EPI images. Depending on this value, the mask will be
-        computed from masking.compute_background_mask or
-        masking.compute_epi_mask.
+    %(mask_strategy)s
 
         .. note::
             This parameter will be ignored if a mask image is provided.
 
+        .. note::
+            Depending on this value, the mask will be computed from
+            :func:`nilearn.masking.compute_background_mask`,
+            :func:`nilearn.masking.compute_epi_mask`, or
+            :func:`nilearn.masking.compute_brain_mask`.
+
+        Default is 'background'.
     %(memory)s
     %(memory_level)s
     %(n_jobs)s